+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - Added DateRangePicker component
 - Added DateTimePicker component
 - Added TimePicker component
+- New Dropdown component
 
 ## 1.0.0 - 06-09-2020